refactor: move tokens modal
