+++ src/components/BucketManager.vue
@@ -42,14 +42,12 @@
       </q-item>
       <q-item>
         <q-item-section>
-          <router-link to="/move-tokens" style="text-decoration: none">
-            <q-btn color="primary" outline>
-              {{ $t("BucketDetail.move") }}
-              <q-tooltip>{{
-                $t("BucketManager.tooltips.move_button")
-              }}</q-tooltip>
-            </q-btn>
-          </router-link>
+          <q-btn color="primary" outline @click="moveTokensOpen = true">
+            {{ $t("BucketDetail.move") }}
+            <q-tooltip>{{
+              $t("BucketManager.tooltips.move_button")
+            }}</q-tooltip>
+          </q-btn>
         </q-item-section>
       </q-item>
     </q-list>
@@ -77,6 +75,7 @@
   <BucketDialog v-model="dialogOpen" />
   <EditBucketModal v-model="editModalOpen" @save="handleEditSave" :bucket="editBucket" />
   <BucketDetailModal v-model="detailModalOpen" :bucket-id="detailBucketId" />
+  <MoveTokensModal v-model="moveTokensOpen" />
 </template>
 
 <script lang="ts">
@@ -91,10 +90,17 @@ import BucketCard from './BucketCard.vue';
 import BucketDialog from './BucketDialog.vue';
 import EditBucketModal from './EditBucketModal.vue';
 import BucketDetailModal from './BucketDetailModal.vue';
+import MoveTokensModal from './MoveTokensModal.vue';
 
 export default defineComponent({
   name: 'BucketManager',
-  components: { BucketCard, BucketDialog, EditBucketModal, BucketDetailModal },
+  components: {
+    BucketCard,
+    BucketDialog,
+    EditBucketModal,
+    BucketDetailModal,
+    MoveTokensModal,
+  },
   setup() {
     const bucketsStore = useBucketsStore();
     const uiStore = useUiStore();
@@ -106,6 +112,7 @@ export default defineComponent({
 
     const editModalOpen = ref(false);
     const detailModalOpen = ref(false);
+    const moveTokensOpen = ref(false);
     const editBucket = ref<any>(null);
     const detailBucketId = ref<string | null>(null);
 
@@ -207,6 +214,7 @@ export default defineComponent({
       showDelete,
       editModalOpen,
       detailModalOpen,
+      moveTokensOpen,
       editBucket,
       detailBucketId,
       openAdd,
